refactor(channels): rename receive.ts → inbound.ts for symmetry with outbound.ts
Mixed naming style (receive verb vs outbound noun) was inconsistent.
inbound/outbound is the standard messaging-domain pair.

Renames:
  receive.ts        → inbound.ts
  receive.test.ts   → inbound.test.ts
  ReceiveDeps       → InboundDeps
  ReceiveCallbacks  → InboundCallbacks
  ReceiveContext    → InboundContext
  GuildReceiveConfig → GuildInboundConfig

Function name receiveDiscordMessage stays — verb-form action name reads
naturally at the call site, and 'handleInbound' would collide with the
existing wrapper in index.ts.
